Fix #15.
The "(Show details)" link is added to the front of the parent element,
which as it is is the label (which contains the radio input, too). There
needs to be an element around just the nomination HTML to do this right.

The awkwardness of blockquotes getting inserted into inline elements
bugs me. I may rethink how we do the nomination HTML somehow.